+++ contracts/counter/tests/checksum.rs
@@ -1,5 +1,7 @@
 use counter_contract::CounterContract;
-use cw_orch::{contract::interface_traits::Uploadable, mock::Mock};
+use cw_orch::{
+    contract::interface_traits::Uploadable, mock::Mock, osmosis_test_tube::MOCK_CHAIN_INFO,
+};
 
 #[test]
 fn checksum() {
@@ -12,12 +14,11 @@ fn checksum() {
     let lines = io::BufReader::new(file).lines();
     let mut found = false;
 
-    for line in lines.flatten() {
+    for line in lines.map_while(Result::ok) {
         if line.contains("counter_contract.wasm") {
             let parts: Vec<&str> = line.split_whitespace().collect();
             if parts.len() > 1 {
-                let calculated_hash = CounterContract::new(Mock::new("sender"))
-                    .wasm()
+                let calculated_hash = CounterContract::<Mock>::wasm(&MOCK_CHAIN_INFO.into())
                     .checksum()
                     .unwrap();
                 assert_eq!(parts[0], calculated_hash.to_string());