+++ cw-orch-daemon/src/state.rs
@@ -1,14 +1,15 @@
 use super::error::DaemonError;
-use crate::{channel::GrpcChannel, networks::ChainKind};
+use crate::{
+    channel::GrpcChannel, env::default_state_folder, env::DaemonEnvVars, networks::ChainKind,
+};
 
 use cosmwasm_std::Addr;
+use cw_orch_core::environment::ChainInfoOwned;
 use cw_orch_core::{
-    env::default_state_folder,
     environment::StateInterface,
     log::{connectivity_target, local_target},
-    CwEnvError, CwOrchEnvVars,
+    CwEnvError,
 };
-use ibc_chain_registry::chain::ChainData;
 use serde::Serialize;
 use serde_json::{json, Value};
 use std::{collections::HashMap, fs::File, path::Path};
@@ -25,7 +26,7 @@ pub struct DaemonState {
     /// gRPC channel
     pub grpc_channel: Channel,
     /// Information about the chain
-    pub chain_data: ChainData,
+    pub chain_data: ChainInfoOwned,
     /// Flag to set the daemon state readonly and not pollute the env file
     pub read_only: bool,
 }
@@ -34,19 +35,19 @@ impl DaemonState {
     /// Creates a new state from the given chain data and deployment id.
     /// Attempts to connect to any of the provided gRPC endpoints.
     pub async fn new(
-        mut chain_data: ChainData,
+        chain_data: ChainInfoOwned,
         deployment_id: String,
         read_only: bool,
     ) -> Result<DaemonState, DaemonError> {
-        if chain_data.apis.grpc.is_empty() {
+        if chain_data.grpc_urls.is_empty() {
             return Err(DaemonError::GRPCListIsEmpty);
         }
 
-        log::debug!(target: &connectivity_target(), "Found {} gRPC endpoints", chain_data.apis.grpc.len());
+        log::debug!(target: &connectivity_target(), "Found {} gRPC endpoints", chain_data.grpc_urls.len());
 
         // find working grpc channel
         let grpc_channel =
-            GrpcChannel::connect(&chain_data.apis.grpc, chain_data.chain_id.as_str()).await?;
+            GrpcChannel::connect(&chain_data.grpc_urls, chain_data.chain_id.as_str()).await?;
 
         // If the path is relative, we dis-ambiguate it and take the root at $HOME/$CW_ORCH_STATE_FOLDER
         let mut json_file_path = Self::state_file_path()?;
@@ -54,7 +55,7 @@ impl DaemonState {
         log::debug!(target: &local_target(), "Using state file : {}", json_file_path);
 
         // if the network we are connecting is a local kind, add it to the fn
-        if chain_data.network_type == ChainKind::Local.to_string() {
+        if chain_data.kind == ChainKind::Local {
             let name = Path::new(&json_file_path)
                 .file_stem()
                 .unwrap()
@@ -68,20 +69,6 @@ impl DaemonState {
 
             json_file_path = format!("{folder}/{name}_local.json");
         }
-
-        // Try to get the standard fee token (probably shortest denom)
-        let shortest_denom_token = chain_data.fees.fee_tokens.iter().fold(
-            chain_data.fees.fee_tokens[0].clone(),
-            |acc, item| {
-                if item.denom.len() < acc.denom.len() {
-                    item.clone()
-                } else {
-                    acc
-                }
-            },
-        );
-        // set a single fee token
-        chain_data.fees.fee_tokens = vec![shortest_denom_token];
 
         // build daemon state
         let state = DaemonState {
@@ -103,7 +90,7 @@ impl DaemonState {
             crate::json_file::write(
                 &state.json_file_path,
                 &state.chain_data.chain_id.to_string(),
-                &state.chain_data.chain_name,
+                &state.chain_data.network_info.chain_name.to_string(),
                 &state.deployment_id,
             );
         }
@@ -115,7 +102,7 @@ impl DaemonState {
     /// Returns the path of the file where the state of `cw-orchestrator` is stored.
     pub fn state_file_path() -> Result<String, DaemonError> {
         // check if STATE_FILE en var is configured, default to state.json
-        let env_file_path = CwOrchEnvVars::load()?.state_file;
+        let env_file_path = DaemonEnvVars::state_file();
         let state_file_path = if env_file_path.is_relative() {
             // If it's relative, we check if it start with "."
             let first_path_component = env_file_path
@@ -154,7 +141,11 @@ impl DaemonState {
     /// Retrieve a stateful value using the chainId and networkId
     pub fn get(&self, key: &str) -> Result<Value, DaemonError> {
         let json = self.read_state()?;
-        Ok(json[&self.chain_data.chain_name][&self.chain_data.chain_id.to_string()][key].clone())
+        Ok(
+            json[&self.chain_data.network_info.chain_name][&self.chain_data.chain_id.to_string()]
+                [key]
+                .clone(),
+        )
     }
 
     /// Set a stateful value using the chainId and networkId
@@ -170,8 +161,8 @@ impl DaemonState {
 
         let mut json = self.read_state()?;
 
-        json[&self.chain_data.chain_name][&self.chain_data.chain_id.to_string()][key]
-            [contract_id] = json!(value);
+        json[&self.chain_data.network_info.chain_name][&self.chain_data.chain_id.to_string()]
+            [key][contract_id] = json!(value);
 
         serde_json::to_writer_pretty(File::create(&self.json_file_path).unwrap(), &json)?;
         Ok(())
@@ -236,9 +227,7 @@ impl StateInterface for DaemonState {
 pub mod test {
     use std::env;
 
-    use cw_orch_core::env::STATE_FILE_ENV_NAME;
-
-    use crate::DaemonState;
+    use crate::{env::STATE_FILE_ENV_NAME, DaemonState};
 
     #[test]
     fn test_env_variable_state_path() -> anyhow::Result<()> {