+++ cw-orch/tests/interface_macro.rs
@@ -1,5 +1,5 @@
 use cw_orch::{
-    environment::{CwEnv, TxHandler},
+    environment::TxHandler,
     prelude::{ContractWrapper, Uploadable},
 };
 
@@ -22,8 +22,8 @@ use cw_orch::interface;
 )]
 pub struct MockContract;
 
-impl<Chain: CwEnv> Uploadable for MockContract<Chain> {
-    fn wrapper(&self) -> <Mock as TxHandler>::ContractSource {
+impl<Chain> Uploadable for MockContract<Chain> {
+    fn wrapper() -> <Mock as TxHandler>::ContractSource {
         Box::new(
             ContractWrapper::new_with_empty(
                 mock_contract::execute,